+++ zenoh/src/net/protocol/mod.rs
@@ -15,4 +15,3 @@
 pub mod core;
 pub mod io;
 pub mod proto;
-pub mod session;
