--- zenoh/src/net/protocol/session/common/mod.rs
@@ -1,23 +1,0 @@
-//
-// Copyright (c) 2017, 2020 ADLINK Technology Inc.
-//
-// This program and the accompanying materials are made available under the
-// terms of the Eclipse Public License 2.0 which is available at
-// http://www.eclipse.org/legal/epl-2.0, or the Apache License, Version 2.0
-// which is available at https://www.apache.org/licenses/LICENSE-2.0.
-//
-// SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
-//
-// Contributors:
-//   ADLINK zenoh team, <zenoh@adlink-labs.tech>
-//
-pub(crate) mod batch;
-pub(crate) mod conduit;
-pub(crate) mod defragmentation;
-pub(crate) mod pipeline;
-pub(crate) mod seq_num;
-
-use super::core;
-use super::io;
-use super::proto;
-use super::session;
