+++ zenoh/src/net/routing/mod.rs
@@ -21,3 +21,4 @@ pub mod router;
 use super::link;
 use super::protocol;
 use super::runtime;
+use super::transport;
