+++ zenoh/tests/intermittent.rs
@@ -25,7 +25,7 @@ use zenoh::net::protocol::core::{
 };
 use zenoh::net::protocol::io::ZBuf;
 use zenoh::net::protocol::proto::ZenohMessage;
-use zenoh::net::protocol::session::{
+use zenoh::net::transport::{
     DummySessionEventHandler, Session, SessionEventHandler, SessionHandler, SessionManager,
     SessionManagerConfig, SessionManagerConfigUnicast,
 };
