+++ zenoh/tests/shm.rs
@@ -23,8 +23,8 @@ mod tests {
     use zenoh::net::protocol::core::{whatami, Channel, PeerId, Priority, Reliability, ResKey};
     use zenoh::net::protocol::io::{SharedMemoryManager, ZBuf};
     use zenoh::net::protocol::proto::{Data, ZenohBody, ZenohMessage};
-    use zenoh::net::protocol::session::unicast::authenticator::SharedMemoryAuthenticator;
-    use zenoh::net::protocol::session::{
+    use zenoh::net::transport::unicast::authenticator::SharedMemoryAuthenticator;
+    use zenoh::net::transport::{
         Session, SessionEventHandler, SessionHandler, SessionManager, SessionManagerConfig,
         SessionManagerConfigUnicast,
     };
